+++ source/game_sa/Audio/Enums/ePedAudioType.h
