+++ source/game_sa/Tasks/TaskTypes/TaskSimpleSay.h
@@ -2,7 +2,7 @@
 
 #include "TaskSimple.h"
 #include "TaskTimer.h"
-#include <Audio/PedSpeechContexts.h>
+#include <Audio/Enums/PedSpeechContexts.h>
 
 class CEvent;
 class CTaskSimpleSay;
