+++ source/game_sa/Timer.h
@@ -13,8 +13,8 @@ class CTimer {
     * So basically, a timestep is just fraction of the frametime (timestep = frametime / TIMESTEP_PER_SECOND)
     * this timestep is used basically everywhere to calculate physics, etc.
     */
-    static constexpr float TIMESTEP_PER_SECOND = 50.f;                         /// Number of steps/second
-    static constexpr float TIMESTEP_LEN_IN_MS  = 1000.f / TIMESTEP_PER_SECOND; /// How long (in ms) a timestep is
+    static constexpr float TIMESTEP_PER_SECOND = 50.f;                         //!< Number of steps/second
+    static constexpr float TIMESTEP_LEN_IN_MS  = 1000.f / TIMESTEP_PER_SECOND; //!< How long (in ms) a timestep is
 
 public:
     typedef uint64(__cdecl* TimerFunction_t)();
